feat: Add Family component for managing family members
This commit adds the Family component to the src/components directory. The Family component allows users to view and add family members. It fetches the family data from the Firestore database and displays it in a list. Users can also add new family members by entering their names in an input field and clicking the "Add" button. The new family member is then added to the database and displayed in the list. This feature enhances the functionality of the application by providing a way to manage family members.